requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
